renamed githubtest.js to get_issues.js
