+++ lib/server.js
@@ -5,7 +5,7 @@ var session = require('express-session');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 var everyauth = require('everyauth');
-var githubIssues = require('../githubtest');
+var githubIssues = require('../get_issues');
 var githubOrgs = require('../get_orgs');
 
 var config = require('./config');
